pyproject.toml
[project]
name = "morning-mailer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.196.0",
    "google-auth-httplib2>=0.4.0",
    "google-auth-oauthlib>=1.4.0",
    "gunicorn>=26.0.0",
    "huey[backends]>=3.0.0",
    "ipython>=9.13.0",
    "langchain>=1.2.18",
    "langchain-google-genai>=4.2.2",
    "langchain-nvidia-ai-endpoints>=1.3.0",
    "langchain-openai>=1.2.1",
    "langchain-openrouter>=0.2.3",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "rich>=15.0.0",
]
